# docs/swagger.yaml
openapi: 3.0.0
info:
  title: API de produtos
  version: 1.0.0
  description: API para gestão de produtos
servers:
  - url: http://localhost:5000
    description: Servidor backend
paths:
  /produtos:
    get:
      summary: Listar todos os produtos
      responses:
        '200':
          description: Lista de produtos
    post:
      summary: Criar um novo produto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nome, preco, estoque, capa, categoria]
              properties:
                nome:
                  type: string
                preco:
                  type: number
                capa:
                  type: string
                fotos:
                  type: array
                  items:
                    type: string
                estoque:
                  type: integer
                categoria:
                  type: string
                rate:
                  type: number
                descricao:
                  type: string
                detalhes:
                  type: string
                comentario:
                  type: string
      responses:
        '201':
          description: Produto criado com sucesso
        '400':
          description: Dados obrigatórios ausentes
  /produtos/offset/{offset}/{limit}:
    get:
      summary: Listar produtos com offset e limite
      parameters:
        - in: path
          name: offset
          required: true
          schema:
            type: integer
        - in: path
          name: limit
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista paginada de produtos
  /produtos/nome/{nome}:
    get:
      summary: Buscar produtos por nome
      parameters:
        - in: path
          name: nome
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lista de produtos encontrados
  /produtos/categoria/{categoria}:
    get:
      summary: Buscar produtos por categoria
      parameters:
        - in: path
          name: categoria
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lista de produtos por categoria
  /produtos/preco/{preco_min}/{preco_max}:
    get:
      summary: Buscar produtos por faixa de preço
      parameters:
        - in: path
          name: preco_min
          required: true
          schema:
            type: integer
        - in: path
          name: preco_max
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de produtos por preço
  /produtos/id/{id_product}:
    get:
      summary: Obter produto por ID
      parameters:
        - in: path
          name: id_product
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Produto encontrado
        '404':
          description: Produto não encontrado
  /produtos/{id_produto}:
    put:
      summary: Atualizar produto por ID
      parameters:
        - in: path
          name: id_produto
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
                preco:
                  type: number
                capa:
                  type: string
                fotos:
                  type: array
                  items:
                    type: string
                estoque:
                  type: integer
                categoria:
                  type: string
                rate:
                  type: number
                descricao:
                  type: string
                detalhes:
                  type: string
                comentario:
                  type: string
      responses:
        '200':
          description: Produto atualizado
        '404':
          description: Produto não encontrado
    delete:
      summary: Apagar produto por ID
      parameters:
        - in: path
          name: id_produto
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Produto apagado com sucesso
        '404':
          description: Produto não encontrado